Updated README and build script

+++ build.sh
@@ -1,5 +1,17 @@
-cd webapp && npm run build && cd ..
-rm a_main-packr.go
+#!/bin/bash
+
+if [ -e ./a_main-packr.go ]; then
+    rm a_main-packr.go
+fi
 packr
-go build -o dist/neutrino.app/Contents/MacOS/neutrino
-open dist/neutrino.app
+
+name="neutrino"
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    go build -o dist/darwin/$name.app/Contents/MacOS/$name
+elif [[ "$OSTYPE" == "linux-gnu" ]]; then
+    go build -o dist/linux/$name
+elif [[ "$OSTYPE" == "msys" ]]; then
+    go build -ldflags="-H windowsgui" -o dist/windows/$name.exe
+fi
+
+open dist/
